--- src/interim/isExecutable.js
@@ -1,12 +1,0 @@
-const os = require('os');
-const isWin = /^win/.test(os.platform());
-
-function isExecutable(extension, mode) {
-  const ox = !!(((mode << 6) >> 6) & 1);
-  const gx = !!(((mode << 3) >> 6) & 1);
-  const ux = !!((mode >> 6) & 1);
-  const exe = /\.(exe|bat|cmd)$/.test(extension);
-  return isWin ? exe : ux || gx || ox;
-}
-
-module.exports = isExecutable;
