+++ src/interim/parsing/parseDirs.js
@@ -1,8 +1,8 @@
 const fs = require('fs');
 const path = require('path');
 const os = require('os');
-const log = require('./log');
-const sharedConfig = require('../renderer/shared/sharedConfig');
+const log = require('../log');
+const sharedConfig = require('../../renderer/shared/sharedConfig');
 
 let counter = 0;
 const getConfig = () => sharedConfig;
