+++ src/interim/parsing/postParsingHooks/processXdgDesktopFile.js
