+++ src/renderer/shared/sharedStore.js
@@ -1,3 +1,8 @@
-const sharedStore = {};
+const sharedStore = {
+  current: 0,
+  ghost: null,
+  searchItems: [],
+  found: [],
+};
 
 module.exports = sharedStore;
